Add search pkg; move Trie to search.Trie
